padded batch + better tokenization encoding+decoding
