+++ src/llm/__init__.py
@@ -1,4 +1,8 @@
-from .wmt_dataset import get_dataset, get_dataloader
-from .tokenizer import get_tokenizers
+from .wmt_dataset import get_dataloader
+from .tokenizer import get_tokenizers, get_dataset
 
-__all__ = [get_dataset, get_dataloader, get_tokenizers]
+__all__ = [
+    "get_dataset",
+    "get_dataloader",
+    "get_tokenizers",
+]
